rename: web_project.db -> editor.db

+++ resetdb.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
-
-rm -rf Migrations
-rm web_project.db
+if test -d "Migrations"; then
+  rm -rf Migrations
+fi
+if test -f "editor.db"; then
+  rm editor.db
+fi
 dotnet ef migrations add InitialCreate
 dotnet ef database update
